Revert "Added [StackTraceHidden] attribute"
This reverts commit 12b088b9ee1ecfbafb5ab3e790072d8c09dda147.
Unfortunately the attribute is actually matched by type and not just by name like the others. We should reintroduce this once either https://github.com/dotnet/runtime/issues/29681 is done, or when the runtime switches to a name matching for the internal attribute, so we can just use a local copy here.