Always include common files in the NuGet package
When the '$(IncludeContentInPack)' property is false, files specified via '@(None)', '@(Content)' items
are excluded from the NuGet package. Adding '@(PackageFile)' items directly to '%(_PackageFiles)' item
via the following target ensures that they will always be included in the package.

So, Use '%(PackageFile)' item to always include files in the package. By default, they are included in
the root of the package but can be overridden via '%(PackageFile.TargetPath)' metadata.